add console related pages
