--- src/components/CourseManagement/index.js
@@ -1,1 +1,0 @@
-export { default } from "./CourseManagement";
