+++ src/components/CourseSettingForm/CourseSettingForm.js
@@ -1,10 +1,9 @@
 import React from "react";
-
 import { Form, Input, InputNumber, Button } from "antd";
 
 const layout = {
-  labelCol: { span: 4 },
-  wrapperCol: { span: 16 },
+  labelCol: { span: 2 },
+  wrapperCol: { span: 12 },
 };
 
 const validateMessages = {
@@ -19,16 +18,11 @@ const validateMessages = {
 
 export default function CourseSettingForm() {
   const onFinish = (values) => {
-    console.log(values);
+    console.log("完成表單", values);
   };
 
   return (
-    <Form
-      {...layout}
-      name="nest-messages"
-      onFinish={onFinish}
-      validateMessages={validateMessages}
-    >
+    <Form {...layout} onFinish={onFinish} validateMessages={validateMessages}>
       <Form.Item name="name" label="課程名稱" rules={[{ required: true }]}>
         <Input />
       </Form.Item>
@@ -40,9 +34,9 @@ export default function CourseSettingForm() {
         <InputNumber />
       </Form.Item>
       <Form.Item name="introduction" label="課程敘述">
-        <Input.TextArea />
+        <Input.TextArea rows={8} />
       </Form.Item>
-      <Form.Item wrapperCol={{ ...layout.wrapperCol, offset: 4 }}>
+      <Form.Item wrapperCol={{ ...layout.wrapperCol, offset: 2 }}>
         <Button type="primary" htmlType="submit">
           確認
         </Button>
