+++ src/components/CoursesTable/dummyData.js
@@ -1,0 +1,50 @@
+export const dummyData = [
+  {
+    key: "1",
+    id: 1,
+    name: "從零開始學習 HTML",
+    updatedAt: "2020-12-16",
+  },
+  {
+    key: "2",
+    id: 2,
+    name: "從零開始學習 CSS",
+    updatedAt: "2020-12-16",
+  },
+  {
+    key: "3",
+    id: 3,
+    name: "從零開始學習 JavaScript",
+    updatedAt: "2020-12-16",
+  },
+  {
+    key: "4",
+    id: 4,
+    name: "從零開始學習 JavaScript",
+    updatedAt: "2020-12-16",
+  },
+  {
+    key: "5",
+    id: 5,
+    name: "從零開始學習 JavaScript",
+    updatedAt: "2020-12-16",
+  },
+  {
+    key: "6",
+    id: 6,
+    name: "從零開始學習 JavaScript",
+    updatedAt: "2020-12-16",
+  },
+  {
+    key: "7",
+    id: 7,
+    name: "從零開始學習 JavaScript",
+    updatedAt: "2020-12-16",
+  },
+  {
+    key: "8",
+    id: 8,
+    name: "從零開始學習 JavaScript",
+    updatedAt: "2020-12-16",
+  },
+];
