+++ src/components/MembersTable/dummyData.js
@@ -1,0 +1,16 @@
+export const dummyData = [
+  {
+    key: "1",
+    id: 1,
+    email: "abc123@gmail.com",
+    auth: "一般",
+    latestLogin: "2020-12-11",
+  },
+  {
+    key: "2",
+    id: 2,
+    email: "a1239999@gmail.com",
+    auth: "管理員",
+    latestLogin: "2020-12-16",
+  },
+];
