+++ src/containers/Routes.js
@@ -1,28 +1,27 @@
 import React from "react";
 import { Switch, Route } from "react-router-dom";
-import Login from "../Pages/Login";
-import Register from "../Pages/Register";
-import Me from "../Pages/Me";
+import Login from "../pages/Login";
+import Register from "../pages/Register";
+import Me from "../pages/Me";
+import ConsolePage from "../pages/ConsolePage";
 //  引入各分頁
-import Cart from '../components/Cart'
-import CourseList from '../components/CourseList'
-import Console from "../components/Console";
-
+import Cart from "../components/Cart";
+import CourseList from "../components/CourseList";
 
 function Routes() {
   return (
     <Switch>
       <Route exact path="/">
         {/* 首頁(課程列表) */}
-        <CourseList/>
+        <CourseList />
       </Route>
       <Route exact path="/courseList">
         {/* 課程列表 */}
-        <CourseList/>
+        <CourseList />
       </Route>
       <Route exact path="/cart">
         {/* 購物車 */}
-        <Cart/>
+        <Cart />
       </Route>
       <Route exact path="/">
         {/* 我的課程 */}
@@ -33,7 +32,7 @@ function Routes() {
       </Route>
       <Route path="/console">
         {/* 管理後台 */}
-        <Console />
+        <ConsolePage />
       </Route>
       <Route exact path="/register">
         <Register />
